# Cache last visited node in paint_arrange_structs