@janisozaur 2017-07-18
Note: this code is in the library part, it is not enough to copy the executable, you have to have separate libraries, so either preload them or use two different build directories.