@janisozaur 2017-07-18
I'll try timing it on raspi in the evening, I'm quite curious.